edit

+++ ForPractice/42_NetLib_ver1/NetLib_ver1/NetLib_ver1/CSession.cpp
@@ -1,3 +1,3 @@
 #include "CSession.h"
-unordered_map<int, CSession*> g_SessionMap;
+unordered_map<__int64, CSession*> g_SessionMap;
 SRWLOCK g_SessionMapLock;
